requirements.txt
streamlit==1.31.0
youtube-transcript-api>=0.6.2
google-generativeai>=0.3.0
python-dotenv==1.0.0
reportlab==4.0.9
pymongo==4.6.1
requests>=2.25.1
